renamed component_xircuits to tray_library
