+++ src/tray_library/Sidebar.tsx
