+++ src/tray_library/TrayItemWidget.tsx
